Wave 90: 4476 -> 4489 (+13), bytes 90.94% -> 92.22%
Four agents through a cap of 3 (W90-A, W90-B, W90-C matching; W90-P1
porting; W90-T tooling). Thirteen functions closed, twelve of them parks,
nine of those carried across three or more waves. Both builds reproduce
14dd0b22c894865867aff89e8116b2dffae25605 (SPLIT_EXIT=0, ASM_EXIT=0);
proto_check clean. parked 98 -> 86.

Matched: sub_08039188 (fixed by PR #3's K&R callee declaration; preflight
reported it UNPARKED), sub_080627F4, sub_08037FD0, sub_08086A58,
sub_08085B30, sub_080860DC (W90-B); sub_080283E4, sub_08020EDC,
sub_0803D558, sub_080620FC, sub_08062C94, sub_08066874 (W90-C);
sub_0804E7A8 (W90-A).

FIRST: PR #3's renames had broken 46 of the 101 unmatched drafts, and
preflight could not tell. Its parked re-test printed `still-fails` for a
draft that does not compile, so six of the ten closest near-misses looked
like ordinary misses. Every draft's last candidate was snapshotted
(_cand.prepr3.bin) and all 46 were ported BY OFFSET, each proved
byte-identical to its pre-merge compile. The hand check was itself
unsound at first (a failed compile leaves the old _cand.bin, which equals
the snapshot); W90-P1 caught it. `struct KeySt` was RESHAPED, not
renamed: old names compile at new offsets and only a byte comparison
catches it. All 89 drafts in wip/ now compile.

THE LEVER OF THE WAVE: WRITE THE STRUCT MEMBER, NOT A POINTER INTO THE
GLOBAL. Six of the thirteen fell to it: sub_08086A58 and sub_080860DC
(`gUnknown_02027F74.unk04[i]` in place of `q = (u8 *)&g; q += 4`),
sub_080283E4 (gPlaySt/gPlayers members), and three map functions via
PR #3's struct Map (`terrain`/`rowOffset` in place of `p + 0x417a` /
`p + 0x1432`). The mechanism, read off an agbcc -dg dump: gcse makes one
shared pseudo for the global's address, and a pointer local that also
uses it passes its register PREFERENCE to unrelated pseudos (unk37
inherited "preferences: 4" from it in sub_080860DC). A residual that
reads as a bare register tie can be caused by an address local in a
different arm. Every one of these parks had been classified "allocation,
nothing behind it". Several were really this.

THE PERMUTER WAS SCORING THE WRONG OBJECTIVE UNTIL THIS WAVE (see the
preceding tools commit). With stack slots scored, sub_080627F4 (parked
since wave 74 as a pure spill-slot swap; W86 concluded no source route
exists) matched on the first 900 s run: `int u;` before `int t;` plus
`s16 c`. PERM_IGNORE/PERM_PRETEND kept sub_08085B30's load-bearing pin
and found `gUnknown_0200FC50 + (i + 2) * 0x80`. Honest negatives: across
13 runs directed search did not beat undirected; the fixed scorer moved
nothing on sub_08031824, sub_08056638 or sub_08055940 in 2 x 900 s; and
permuter score disagreed with byte count on two functions, so judge
candidates by `bytes:`.

OTHER MECHANISMS, each in docs/agbcc-codegen.md:
- Reload hands out spill registers ROUND-ROBIN in insn order, and jump2's
  `if (foo) bar; else break;` range swap reorders blocks after reload.
  sub_08037FD0 closed by putting the c-arm first with its own fail call
  in each arm; it had been retired as having no source handle.
  Registers on `ldr rN,[sp,#k]` / `mov rN,ip` / `movs rN,#imm` scratch
  are reloads: count them in insn order in a dump first.
- agbcc accepts gcc's -d flags; .loop/.gcse/.greg name the pass that owns
  a residual (W90-A). sub_0807E980 99.0 -> 99.4% from the .loop dump:
  the ROM reduces dst as a giv in loop pass 2, and all 144 spellings
  tried reduce it in pass 1.
- Allocno priority is computable from .greg (floor_log2(refs)*refs /
  live_length); each do/while(0) level adds one ref.
- sub_08066874: goto/label control flow starts a new CSE block, so the
  arm re-derives &G from r8 after the call.
- sub_080620FC (hand draft 29.3%/+4): a struct-wrapper chain fixes the
  order of the offset terms; the old draft had the record's x at +0,
  which was semantically wrong.

Prototype change: sub_08020EDC's 6th parameter u8 -> int (the member
form needed it), with (u8) casts at its only C caller, c_080210C8.c,
re-verified.

Parked, improved: sub_0807E980 99.4% (6 bytes), sub_08054C5C 93.4%.
sub_0805A0EC stays at 2 code bytes, reload round-robin one step past
reach. sub_08031824 moved from asm to parked with its wave-90 record.

+++ include/unknown-functions.h
@@ -10581,16 +10581,16 @@ void sub_080200EC(s16, s16, s16, s16);
  * as `movs #1; rsbs` -- so signed. Byte-neutral there; c_08020D50 re-verified
  * by try_match exit code after the change. */
 void sub_08020B88(s16, s16, s16, s16);
-/* Wave 73, W73-F: the prose above says arguments 5 AND 6 are `int`, which
- * CONTRADICTS this declaration and is the stale half. The 6th is `u8` and the
- * caller settles it: src/decomp/c_080210C8.c is matched, holds its own sixth
- * parameter as `int a6`, and passes it through `lsls #0x18; lsrs #0x18` -- that
- * truncation is the int->u8 conversion this declaration's 6th parameter
- * requires, and it disappears if the parameter is widened, costing the matched
- * caller 4 bytes. decomp-permuter reaches 91.5% on the definition (against
- * 90.7%) with `unsigned int` here; it rewrites the prototype block and so never
- * compiles that against this header. Do not widen it. */
-void sub_08020EDC(s16, s16, s16, u8 *, int, u8);
+/* Wave 90, W90-C: the 6th parameter is `int`, as the prose above says, and
+ * the definition proves it. Its prologue narrows the 6th only at the
+ * `f = flags;` copy, AFTER the `d = delta;` group; a `u8` formal would be
+ * narrowed first by PROMOTE_MODE, which put one `lsls` a slot early and was
+ * the last residual of sub_08020EDC. The matched caller
+ * src/decomp/c_080210C8.c keeps its `lsls #0x18; lsrs #0x18` with an explicit
+ * `(u8)a6` at each call, which is the same tree the old prototype built, and
+ * was re-verified by trymatch exit code. (Wave 73 read that narrowing as
+ * proof of a `u8` formal; a cast at the call site gives it too.) */
+void sub_08020EDC(s16, s16, s16, u8 *, int, int);
 
 /* Wave 36, W36-M. Every signature below is COPIED VERBATIM from a byte-verified
  * definition in src/decomp/ -- none of them had a declaration in any header,
